Improve OpenOS code quality, security, and documentation

+++ Kernel2.0/keyboard.h
@@ -6,6 +6,7 @@
 #define KEYBOARD_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 /* Keyboard I/O port */
 #define KEYBOARD_DATA_PORT 0x60
@@ -18,6 +19,6 @@ void keyboard_init(void);
 void keyboard_handler(void);
 
 /* Get a line of input (blocking) */
-void keyboard_get_line(char* buffer, int max_len);
+void keyboard_get_line(char* buffer, size_t max_len);
 
 #endif /* KEYBOARD_H */
